// sourcecode/BTrak.Packages/snova-audits-module/projects/project-components/src/lib/audits/dependencies/models/projectFeature.ts
export class ProjectFeature {
  projectFeatureId: string;
  projectFeatureName: string;
  projectId: string;
  projectFeatureResponsiblePersonId: string;
  isArchived: boolean;
  IsDelete: boolean;
  projectFeatureResponsiblePersonName: string;
  profileImage: string;
  roleName: string;
  projectName:string;
  projectFeatureResponsiblePerson:any;
  timeStamp: any;
  createdDateTime: any;
 
}